+++ src/main/java/nl/jqno/equalsverifier/Warning.java
@@ -163,13 +163,14 @@ public enum Warning {
     STRICT_INHERITANCE,
 
     /**
-     * Disables the check that fields marked with the @Id annotation in JPA
-     * entities may not be used in the {@code equals} contract.
+     * Disables the check that fields marked with the @Id or @EmbeddedId
+     * annotations in JPA entities may not be used in the {@code equals}
+     * contract.
      * <p>
-     * When this warning is suppressed, the fields marked with @Id will become
-     * the entity's surrogate key. Only these fields can now be part of the
-     * {@code equals} contract; all other fields may no longer be used in
-     * {@code equals}.
+     * When this warning is suppressed, the fields marked with @Id
+     * or @EmbeddedId will become the entity's surrogate key. Only these fields
+     * can now be part of the {@code equals} contract; all other fields may no
+     * longer be used in {@code equals}.
      */
     SURROGATE_KEY,