+++ src/main/java/nl/jqno/equalsverifier/internal/checkers/fieldchecks/SignificantFieldCheck.java
@@ -122,7 +122,7 @@ public class SignificantFieldCheck<T> implements FieldCheck {
         }
         else if (anotherFieldIsMarkedAsId) {
             message = "Significant fields: equals does not use %%, or it is stateless.\n" +
-                "Suppress Warning.SURROGATE_KEY if you want to use only the @Id field(s).";
+                "Suppress Warning.SURROGATE_KEY if you want to use only the @Id or @EmbeddedId field(s).";
         }
         else {
             message = "Significant fields: equals does not use %%, or it is stateless.";
@@ -137,11 +137,11 @@ public class SignificantFieldCheck<T> implements FieldCheck {
         final String message;
         if (thisFieldIsMarkedAsId) {
             message = "Significant fields: %% is marked @Id or @EmbeddedId so equals should not use it, but it does.\n" +
-                "Suppress Warning.SURROGATE_KEY if you want to use only the @Id field(s).";
+                "Suppress Warning.SURROGATE_KEY if you want to use only the @Id or @EmbeddedId field(s).";
         }
         else if (anotherFieldIsMarkedAsId) {
             message = "Significant fields: equals should not use %% because Warning.SURROGATE_KEY is suppressed" +
-                " and it is not marked as @Id, but it does.";
+                " and it is not marked as @Id or @EmbeddedId, but it does.";
         }
         else {
             message = "Significant fields: equals should not use %%, but it does.";